% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mq_service.R
\name{mq}
\alias{mq}
\title{AmazonMQ}
\usage{
mq()
}
\description{
Amazon MQ is a managed message broker service for Apache ActiveMQ that makes it easy to set up and operate message brokers in the cloud. A message broker allows software applications and components to communicate using various programming languages, operating systems, and formal messaging protocols.
}
\section{Operations}{

\tabular{ll}{
\link[=mq_create_broker]{create_broker} \tab Creates a broker \cr
\link[=mq_create_configuration]{create_configuration} \tab Creates a new configuration for the specified configuration name \cr
\link[=mq_create_tags]{create_tags} \tab Add a tag to a resource \cr
\link[=mq_create_user]{create_user} \tab Creates an ActiveMQ user \cr
\link[=mq_delete_broker]{delete_broker} \tab Deletes a broker \cr
\link[=mq_delete_tags]{delete_tags} \tab Remove a tag from a resource \cr
\link[=mq_delete_user]{delete_user} \tab Deletes an ActiveMQ user \cr
\link[=mq_describe_broker]{describe_broker} \tab Returns information about the specified broker \cr
\link[=mq_describe_configuration]{describe_configuration} \tab Returns information about the specified configuration \cr
\link[=mq_describe_configuration_revision]{describe_configuration_revision} \tab Returns the specified configuration revision for the specified configuration\cr
\link[=mq_describe_user]{describe_user} \tab Returns information about an ActiveMQ user \cr
\link[=mq_list_brokers]{list_brokers} \tab Returns a list of all brokers \cr
\link[=mq_list_configuration_revisions]{list_configuration_revisions} \tab Returns a list of all revisions for the specified configuration \cr
\link[=mq_list_configurations]{list_configurations} \tab Returns a list of all configurations \cr
\link[=mq_list_tags]{list_tags} \tab Lists tags for a resource \cr
\link[=mq_list_users]{list_users} \tab Returns a list of all ActiveMQ users \cr
\link[=mq_reboot_broker]{reboot_broker} \tab Reboots a broker \cr
\link[=mq_update_broker]{update_broker} \tab Adds a pending configuration change to a broker \cr
\link[=mq_update_configuration]{update_configuration} \tab Updates the specified configuration \cr
\link[=mq_update_user]{update_user} \tab Updates the information for an ActiveMQ user
}
}

\examples{
\donttest{svc <- mq()
svc$create_broker(
  Foo = 123
)}

}
