% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mq_operations.R
\name{mq_describe_configuration}
\alias{mq_describe_configuration}
\title{Returns information about the specified configuration}
\usage{
mq_describe_configuration(ConfigurationId)
}
\arguments{
\item{ConfigurationId}{[required] The unique ID that Amazon MQ generates for the configuration.}
}
\description{
Returns information about the specified configuration.
}
\section{Request syntax}{
\preformatted{svc$describe_configuration(
  ConfigurationId = "string"
)
}
}

\keyword{internal}
