% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mq_operations.R
\name{mq_list_tags}
\alias{mq_list_tags}
\title{Lists tags for a resource}
\usage{
mq_list_tags(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] the Amazon Resource Name (ARN)}
}
\description{
Lists tags for a resource.
}
\section{Request syntax}{
\preformatted{mq$list_tags(
  ResourceArn = "string"
)
}
}

\keyword{internal}
