% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_operations.R
\name{sfn_describe_activity}
\alias{sfn_describe_activity}
\title{Describes an activity}
\usage{
sfn_describe_activity(activityArn)
}
\arguments{
\item{activityArn}{[required] The Amazon Resource Name (ARN) of the activity to describe.}
}
\description{
Describes an activity.
}
\details{
This operation is eventually consistent. The results are best effort and
may not reflect very recent updates and changes.
}
\section{Request syntax}{
\preformatted{svc$describe_activity(
  activityArn = "string"
)
}
}

\keyword{internal}
