% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_operations.R
\name{sfn_send_task_success}
\alias{sfn_send_task_success}
\title{Used by workers to report that the task identified by the taskToken
completed successfully}
\usage{
sfn_send_task_success(taskToken, output)
}
\arguments{
\item{taskToken}{[required] The token that represents this task. Task tokens are generated by the
service when the tasks are assigned to a worker (see
GetActivityTaskOutput\$taskToken).}

\item{output}{[required] The JSON output of the task.}
}
\description{
Used by workers to report that the task identified by the \code{taskToken}
completed successfully.
}
\section{Request syntax}{
\preformatted{svc$send_task_success(
  taskToken = "string",
  output = "string"
)
}
}

\keyword{internal}
